requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.0
django-multiselectfield==0.1.12
h11==0.14.0
html5lib==1.1
idna==3.4
lxml==4.9.2
outcome==1.2.0
PySocks==1.7.1
pytz==2022.7.1
requests==2.28.2
selenium==4.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sqlparse==0.4.3
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.14
webencodings==0.5.1
wsproto==1.2.0
